backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.7.4
pydantic-settings>=2.1.0

# LangChain & RAG
langchain
langchain-google-genai
langchain-community
langchain-qdrant

# Vector Database
qdrant-client

# Google AI
google-generativeai

# Reranker
requests

# Text Processing
tiktoken
pypdf
python-docx

# Environment
python-dotenv
